Initial builder completion
