+++ gneiss-mqtt/src/operation/mod.rs
@@ -13,6 +13,7 @@ use crate::*;
 use crate::alias::*;
 use crate::client::*;
 use crate::client::shared_impl::*;
+use crate::config::*;
 use crate::decode::*;
 use crate::encode::*;
 use crate::encode::utils::*;
